requirements.txt
asgiref==3.5.2
attrs==22.2.0
black==22.3.0
click==8.1.3
Django==4.0.4
django-extensions==3.1.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
exceptiongroup==1.1.0
Faker==13.12.0
iniconfig==2.0.0
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
pytest==7.2.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
